Re-organize views into more logical groups
